+++ packages/web-frontend/public/index.html
@@ -159,6 +159,9 @@
     <noscript>
       <p style="color: white">You need to enable JavaScript to run this app.</p>
     </noscript>
+    <!-- App wide loading spinner.
+      It is placed in the index.html file so that it displays while the react app is mounting and it
+      is outside of the root so that it still displays while we fetch essential initial app data (eg. landing pages)-->
     <div id="tupaia-spinner"></div>
     <div id="root"></div>
   </body>
