+++ packages/web-frontend/src/screens/desktop/RootScreen/MainPage.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
@@ -50,8 +50,21 @@ const MapContainer = styled.div`
 `;
 
 const MainPage = ({ enlargedDialogIsVisible, isLoading, sidePanelWidth }) => {
-  const { isCustomLandingPage } = useCustomLandingPages();
-  if (isCustomLandingPage) return <LandingPage />;
+  const { isCustomLandingPageLoading, isCustomLandingPage } = useCustomLandingPages();
+
+  // Once the custom landing pages are loaded, hide the tupaia app loader in index.html
+  useEffect(() => {
+    if (isCustomLandingPageLoading === false) {
+      const el = document.getElementById('tupaia-spinner');
+      if (el) {
+        el.remove();
+      }
+    }
+  }, [isCustomLandingPageLoading]);
+
+  if (isCustomLandingPage) {
+    return <LandingPage />;
+  }
 
   return (
     <>
